VerilatorBackend: verilatorScript.sh generate multiline command output
Previously on windows it would generate with \r\n EOL and replace the \n
with a space, which would work ok under sbt/IDE but if you tried
to run the script on command line it would error.

Now we remove all \r but also to help make the script more readable use
the shell backslash continuation character and keep the multiline output.
Now when you want to continue the same command start with at least one
space in the left gutter, but when you want a new command start it flush
with gutter.

Also added "$@" sequence on the end to allow more options to be provided
manually during exec or from command line with diagnosing.