Formal Unit test result log output and diagnostic changes
This better informs the log output with the situation concerning
Formal unit testing.  Previously there would be this alarming naked
Exception("SymbiYosys failure") which occurrs if the exitStatus of
the sby process was non-zero, even if the context driving it was wrapped
in shouldFail({}).

This adds additional parsing of the sby output logs to confirm the task
completed and the logfile is complete and provides a result verdict.

The default behaviour to cleanup() has been changed, with a new option
withoutDebug() has been provided to allow cleanup.